app updates for new bootstrap
